Fix: unordered device properties
This fix comes with a trade-off: it removes the uniqueness constraint on *deprecated* and *debugconfig* elements. But it seems more acceptable to relax this constraint than to impose one on the order, which is not compatible with the specification. What's more, the uniqueness constraint was in any case not imposed by the schema on *sequences* elements, although they are also subject to this constraint in the specification. These uniqueness constraints can instead be verified by the parser.